fixed64: codemod pass 5 -- constructor/function Fixed args + enum consts
- Extended the arg-caster to vector/constructor calls: FVector2/3(fixed,
  fixed) and similar get (double) casts on their Fixed args (the ctors
  take double; same implicit conversion as before). 44 casts.
- d_dehacked MBF_BOUNCE_* enum values compute in pure int ((1<<FRACBITS))
  since FRACUNIT is now a non-integral-constant Fixed.
Behavior-preserving; flag-off build green; tests green.

+++ src/zandronum/src/p_glnodes.cpp
@@ -143,7 +143,7 @@ static int CheckForMissingSegs()
 		if (seg->sidedef!=NULL)
 		{
 			// check all the segs and calculate the length they occupy on their sidedef
-			TVector2<double> vec1(seg->v2->x - seg->v1->x, seg->v2->y - seg->v1->y);
+			TVector2<double> vec1((double)(seg->v2->x - seg->v1->x),(double)( seg->v2->y - seg->v1->y));
 			added_seglen[seg->sidedef - sides] += float(vec1.Length());
 		}
 	}
@@ -153,7 +153,7 @@ static int CheckForMissingSegs()
 		side_t * side =&sides[i];
 		line_t * line = side->linedef;
 
-		TVector2<double> lvec(line->dx, line->dy);
+		TVector2<double> lvec((double)(line->dx),(double)( line->dy));
 		float linelen = float(lvec.Length());
 
 		missing += (added_seglen[i] < linelen - FRACUNIT);
