+++ src/zandronum/src/p_teleport.cpp
@@ -128,7 +128,7 @@ bool P_Teleport (AActor *thing, fixed_t x, fixed_t y, fixed_t z, angle_t angle,
 	ceilingheight = destsect->ceilingplane.ZatPoint (x, y);
 	if (thing->flags & MF_MISSILE)
 	{ // We don't measure z velocity, because it doesn't change.
-		missilespeed = xs_CRoundToInt(TVector2<double>(thing->velx, thing->vely).Length());
+		missilespeed = xs_CRoundToInt(TVector2<double>((double)(thing->velx),(double)( thing->vely)).Length());
 	}
 	if (keepHeight)
 	{
